Use trailing hashes to mark a header as "not important for linking"
This might allow us to remove some of the duplications for some keywords.
In 90% of all cases `data` should link to the variable type for containers,
not to this attribute in user promises. So ignore this - linking to this
particular section then requires an explicit link to be created.

Needs work on the generator.