Change: Add note about implications of intermittent report collection.
This is an important concept. It identifies how we protect from data
loss in the event of a client not being able to be collected from for a
period of time as well as highlighting the implications of this occuring
for a larger population.

(cherry picked from commit e7233f0a7c5f6c22b4c392a11e42fa66f6666be6)