Change: Fix link to old bucket
We don't need to preserve this outdated instruction just because a video
shows that URL. It is better to have up to date text, and not have the
quickinstall scripts duplicated in multiple places. Though we will
continue to publish the quickinstall script at the old location as well
for some time.