Enable the generator to detect the correct lib dir from the branch name
This means you have to pass in the branch name when you start the
generator. The master branch is a special case in which case it will
select the lexically last directory that starts with a number (so for
example 3.7 over 3.6).

This change was necessary after masterfiles received reorganization
which put some key files in versioned directories.