Redmine #7371: Clarify actions when upgrading hub.
The daemons should be restarted from scratch, because some may be
running but others not, due to the policy being invalid after upgrade.
Typically this might happen after upgrading to a new major version
which triggers CFEngine to look inside a differently versioned
directory inside inputs/lib, which may not exist.