CFE-1120: Improve docs related to detecting if files are symlinks
The fileexists() function tells you if a file exists, if it is checking
a broken symlink, that broken symlink exists and therefore the function
returns true.

Checking if a symlink is broken or if it resolves to a specific target
can be accomplished using filestat() combined with fileexists or by
calling out to test binary with returnszero().

(cherry picked from commit e0ddd80313f12c5decda6444d7d9b3a94b63f305)