Fix documentation for audit in action bodies
This attribute was removed for 3.6.0 as part of a massive cleanup. Related
attributes that were removed at the same time include value_kept, value_notkept,
and value_repaired.

They were intended to allow the encoding of business value into the policy, but
were removed to improve maintainability, and because at the time, there was a
low level of utilization, finally, similar data can be tracked in other metadata
like meta tags, comments, and promisees/stakeholders.