Make meta attribute description less confusing
It was very confusingly circular.  You look up "meta attribute"
and you would find, "Description: A list of meta attributes."

That might be amusingly self-referential but it's horrible pedagogy.

Since "attribute" is THE word used for the things we put on promises
(i.e. "promises have attributes"), it is extremely confusing to also
refer to the RHS of a particular attribute as "containing attributes."

Further, this is the only place in the documentation where these are
even *called* attributes.  Everywhere else, they are referred to as
"meta tags" or more often simply "tags".  But you won't find a
definition for "tags" anywhere prior to this commit; it just had to
be figured out by context and by the examples.