Published special topics Security
I made some tweaks which should help pass the build, and I added a comment with
the text from the old cfengine.com/learn section on security. The comment
includes the link to the youtube video presentation where Diego discusses
security. Since the video is private, I figured that this was the best place to
leave a bread crumb.

Ticket: ENT-6853
Changelog: None