Added a section about MPF to the beginning of the Augments documentation
In my experience, it is very common for people to encounter the
Augments file format for the first time when they want to
interact with the MPF, commonly by changing some value,
or to add bundles and inputs.

The Augments documentation was not very helpful in this aspect,
it barely mentioned the MPF or any of the "most common" ways to
use Augments.

To fix this I added a short section with an explanation and
example to the beginning, and a link to the MPF documentation,
which is maybe what a lot of people actually need.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit cdda5bbe7fe38cf30116ade2a65eaae3b777c775)