Added vagrant specific package build version
3.20.0 was one of those releases where all the stars were not quite aligned. We
have 3.20.0-2 hub packages and when we manually fixed up the vagrant tarball we
gave that a release of 2 as well. Unfortunately, that doesn't align with what we
have in the docs where we reference a single version.

So, this separation will allow us to set the build for the vagrant package separately.