Fixed capitalization of headings to be sentence case
A lot of inconsistencies here as well. There's likely
more left, since it's a semi manual process, but it's
easier for me to do it in a few steps.

Found headings with multiple capital letters with bespoke
regexes, like this:

```
^##+ .*[A-Z].*[a-z].*[A-Z][a-z].*
```

I then put all the replacements in a text file and used a
python script to do the actual replacements, so I can replay
it in another branch.

Ticket: ENT-9964
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>