Fixed host ownership after container builds
Inside of the container build we chown all files to jenkins:jenkins but this results
in the host having incorrect ownership after the run which then causes problems
when CI comes back again to checkout with git.

Fix is to chown back to the local uid/gid regardless of outcome of run.sh (the build).

Ticket: ENT-10474
Changelog: none