Removed incorrect claim that `arglist` attribute preserve spaces
Removed incorrect claim that the `arglist` attribute of the `commands`
promise type, preserves spaces in the arguments. I believe it is meant
to be that way, but it currently isn't. Thus we can revert this commit
from master, once it is implemented in one of the next minor releases.
For patch releases, we should keep the behaviour as is, since it may be
relied upon.

Ticket: CFE-2724
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>