markdowner.py: Improved regex replacement log message and search
Put what we are actually replacing in the log message, not
the regex pattern we searched for.

Limit the times we use regex search, when we find a match,
we can just use that instead of doing re.sub().

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 3f476fb8606405e2519cc5e3abf90e3759ca0476)