Changed a couple links to GitHub to use search
We don't actually branch Masterfiles for every minor release, only for LTS
releases. So, targeting the .x path would not work for example in 3.25 since
there is only a 3.25.0 tag and there is no 3.25.x branch.

This switches to use the search to find the same files in the repository. It's
less specific, but more resilient to breakage I think.

(cherry picked from commit 6ce488d6320e41864692e99a33ee15667ca43d61)